Rename standard -> general
